pr_msgs: add generic SignalAction
The SignalAction is an action which does not take any parameters, does not
provide intermediate feedback, and returns no result besides the Goal being
finished or aborted. It is used to trigger actions and events, like stopping.
Triggered actions might also just be a service call. We want to wrap it in
an action to have a unified interface and perspective from the skiller.